feat: Renamed 'taito o-' prefix to 'taito oper-'

+++ plugins/docker/oper-restart.sh
@@ -1,11 +1,13 @@
 #!/bin/bash
 
 : "${taito_cli_path:?}"
+: "${taito_plugin_path:?}"
 
 echo
-echo "### docker - o-stop: Stopping ###"
+echo "### docker - oper-restart: Restarting ###"
 
 "${taito_cli_path}/util/execute-on-host-fg.sh" "docker-compose down" && \
+echo "${taito_plugin_path}/util/start.sh" "${@}" && \
 
 # Call next command on command chain
 "${taito_cli_path}/util/call-next.sh" "${@}"
